Display externalPortals in readmePage
