+++ editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Search/ExternalPortalList/types.ts
@@ -1,0 +1,9 @@
+export interface ExternalPortalsProps {
+  externalPortals: Record<
+    string,
+    {
+      name: string;
+      href: string;
+    }
+  >;
+}
